+++ source/databricks/tests/test_time_series_transformer.py
@@ -24,7 +24,7 @@ from pyspark.sql.types import (
 )
 from datetime import datetime
 from decimal import Decimal
-from package.transforms.time_series_transformer import (
+from package.transformations.time_series_transformer import (
     transform_unprocessed_time_series_to_points,
 )
 from package.codelists import Resolution
